requirements.txt
streamlit==1.24.0
tensorflow==2.13.0
numpy==1.23.5
pillow==9.5.0
matplotlib
typing-extensions==4.5.0